+++ src/com/bjtu/model/pojo/Tb_file.java
@@ -24,7 +24,7 @@ public class Tb_file {
 	//文件的MD5数值，文件夹为""
 	private String md5;
 	//文件的大小，单位是MB
-	private float size_mb;
+	private double size_mb;
 	//文件的大小，单位是B
 	private String size_b;
 	
@@ -95,10 +95,10 @@ public class Tb_file {
 	public void setMd5(String md5) {
 		this.md5 = md5;
 	}
-	public float getSize_mb() {
+	public double getSize_mb() {
 		return size_mb;
 	}
-	public void setSize_mb(float size_mb) {
+	public void setSize_mb(double size_mb) {
 		this.size_mb = size_mb;
 	}
 	public String getSize_b() {
